Add checks before creating/dropping index in migration

+++ src/Migrations/Version20260227151801.php
@@ -15,12 +15,16 @@ final class Version20260227151801 extends AbstractMigration
     #[\Override]
     public function up(Schema $schema): void
     {
-        $this->addSql('CREATE UNIQUE INDEX channel_customer_idx ON webgriffe_sylius_active_campaign_channel_customer (channel_id, customer_id)');
+        if (!$schema->getTable('webgriffe_sylius_active_campaign_channel_customer')->hasIndex('channel_customer_idx')) {
+            $this->addSql('CREATE UNIQUE INDEX channel_customer_idx ON webgriffe_sylius_active_campaign_channel_customer (channel_id, customer_id)');
+        }
     }
 
     #[\Override]
     public function down(Schema $schema): void
     {
-        $this->addSql('DROP INDEX channel_customer_idx ON webgriffe_sylius_active_campaign_channel_customer');
+        if ($schema->getTable('webgriffe_sylius_active_campaign_channel_customer')->hasIndex('channel_customer_idx')) {
+            $this->addSql('DROP INDEX channel_customer_idx ON webgriffe_sylius_active_campaign_channel_customer');
+        }
     }
 }
